+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `/blocks/latest/txs/cbor` endpoint
+
 ## [3.1.2] - 2025-01-24
 
 ### Fixed